structure updates, added beep.py
Bringing set_weight out of the node object.  That way we can use it as a
regular function, since it does not make sense for the function to be
dependent on the node itself.

beep.py imported, will use it during testing to notify me when things
are done loading (this does take a while)